+++ NEWS.md
@@ -16,6 +16,7 @@ Version 1.7.4.9000
 * [#778](https://github.com/tidyverse/lubridate/issues/778) `duration()` works with repeated units
 * [#682](https://github.com/tidyverse/lubridate/issues/682) Fix quarter extraction with small `fiscal_start`s.
 * [#703](https://github.com/tidyverse/lubridate/issues/703) `leap_year()` works with objects supported by `year()`.
+* [#732](https://github.com/tidyverse/lubridate/issues/732) addition works with infinite dates and timespans.
 
 
 Version 1.7.4